backend/views.py
from rest_framework import generics
from backend.models import Fact
from backend.serializers import FactListCreateSerializer


class FactListCreateAPI(generics.ListCreateAPIView):
    queryset = Fact.objects.all()
    serializer_class = FactListCreateSerializer


class FactRetrieveUpdateDestroyAPIView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Fact.objects.all()
    serializer_class = FactListCreateSerializer
