requirements.txt
asgiref==3.4.1
backcall==0.2.0
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
decorator==5.1.1
Django==3.2
django-environ==0.8.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
ipdb==0.13.9
ipython==7.16.3
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
traitlets==4.3.3
typed-ast==1.5.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.6.0
